Rename socialHeaders file as utils
